Rm transactions and UserSequence + some bugfixes

+++ Services/DeleteMessageJsonRpc.cs
@@ -11,8 +11,14 @@ public class DeleteMessageJsonRpc : IJsonRpcHandler<DeleteMessageRequest, EmptyR
         _journalService = journalService;
     }
 
-    public async Task<EmptyResponse> Execute(DeleteMessageRequest request)
+    public async Task<EmptyResponse> Execute(ulong userId, DeleteMessageRequest request)
     {
+        if (request.Id.UserId != userId)
+        {
+            throw new Exception(
+                $"You cannot delete something that is not yours. UserId {userId}, MessageId.UserId {request.Id.UserId}");
+        }
+        
         await _journalService.DeleteMessage(request.Id);
         return new EmptyResponse();
     }
